[Fix] 좋아요 수 Long 타입으로 변경 배포

+++ src/main/java/org/winey/server/service/FeedLikeService.java
@@ -44,6 +44,6 @@ public class FeedLikeService {
             feedLikeRepository.deleteByFeedAndUser(feed, user);
         }
 
-        return CreateFeedLikeResponseDto.of(feedId, feedLike, feedLikeRepository.countByFeed(feed));
+        return CreateFeedLikeResponseDto.of(feedId, feedLike, (long) feedLikeRepository.countByFeed(feed));
     }
 }
